Fix Equal() and GetHashCode()
