+++ src/neo-vm/Types/CompoundType.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 
@@ -21,16 +20,6 @@ namespace Neo.VM.Types
         internal abstract int SubItemsCount { get; }
 
         public abstract void Clear();
-
-        public override bool Equals(object obj)
-        {
-            return ReferenceEquals(this, obj);
-        }
-
-        public override int GetHashCode()
-        {
-            throw new NotSupportedException();
-        }
 
         public override bool ToBoolean()
         {
