+++ src/neo-vm/Types/InteropInterface.cs
@@ -15,16 +15,16 @@ namespace Neo.VM.Types
             _object = value ?? throw new ArgumentException();
         }
 
-        public override bool Equals(object obj)
+        public override bool Equals(StackItem other)
         {
-            if (ReferenceEquals(this, obj)) return true;
-            if (obj is InteropInterface i) return _object.Equals(i._object);
+            if (ReferenceEquals(this, other)) return true;
+            if (other is InteropInterface i) return _object.Equals(i._object);
             return false;
         }
 
         public override int GetHashCode()
         {
-            throw new NotSupportedException();
+            return HashCode.Combine(_object);
         }
 
         public T GetInterface<T>()
